stellar/tools/pyproject.toml
[tool.poetry]
name = "stellar"
version = "0.1.9"
description = "Python-based CLI tool to manage stellar Gateway."
authors = ["Katanemo Labs, Inc."]
packages = [
    { include = "cli" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
stellar_modelserver = "^0.1.9"
click = "^8.1.7"
jinja2 = "^3.1.4"
jsonschema = "^4.23.0"
setuptools = "75.5.0"
docker = "^7.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
stellar = "cli.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
